@@ -1,6 +1,6 @@
 {
   "name": "webpack-chain",
-  "version": "6.5.1",
+  "version": "7.0.0",
   "main": "src/Config.js",
   "typings": "types/index.d.ts",
   "repository": "neutrinojs/webpack-chain",
@@ -38,8 +38,9 @@
   "devDependencies": {
     "@types/enhanced-resolve": "^3.0.6",
     "@types/tapable": "^1.0.6",
-    "@types/webpack": "^4.41.21",
+    "@types/webpack": "^5.0.0",
     "auto-changelog": "^2.2.0",
+    "enhanced-resolve": "^5.8.2",
     "eslint": "^7.5.0",
     "eslint-config-airbnb-base": "^14.2.0",
     "eslint-config-prettier": "^8.0.0",
@@ -48,6 +49,6 @@
     "jest": "^27.0.0",
     "prettier": "^2.0.5",
     "typescript": "^4.0.0",
-    "webpack": "^4.43.0"
+    "webpack": "^5.0.0"
   }
 }
