@@ -1,4 +1,5 @@
 const ChainedMap = require('./ChainedMap');
+const ChainedValueMap = require('./ChainedValueMap');
 const ChainedSet = require('./ChainedSet');
 const Resolve = require('./Resolve');
 const ResolveLoader = require('./ResolveLoader');
@@ -12,35 +13,61 @@ const Performance = require('./Performance');
 module.exports = class extends ChainedMap {
   constructor() {
     super();
-    this.devServer = new DevServer(this);
+    // https://webpack.js.org/configuration/entry-context/
     this.entryPoints = new ChainedMap(this);
-    this.module = new Module(this);
-    this.node = new ChainedMap(this);
-    this.optimization = new Optimization(this);
+    // https://webpack.js.org/configuration/output/
     this.output = new Output(this);
-    this.performance = new Performance(this);
-    this.plugins = new ChainedMap(this);
+    // https://webpack.js.org/configuration/module/
+    this.module = new Module(this);
+    // https://webpack.js.org/configuration/resolve
     this.resolve = new Resolve(this);
+    // https://webpack.js.org/configuration/resolve/#resolveloader
     this.resolveLoader = new ResolveLoader(this);
+    // https://webpack.js.org/configuration/optimization/
+    this.optimization = new Optimization(this);
+    // https://webpack.js.org/configuration/plugins/
+    this.plugins = new ChainedMap(this);
+    // https://webpack.js.org/configuration/dev-server/
+    this.devServer = new DevServer(this);
+    // https://webpack.js.org/configuration/performance/
+    this.performance = new Performance(this);
+    // https://webpack.js.org/configuration/node/
+    this.node = new ChainedValueMap(this);
     this.extend([
+      // https://webpack.js.org/configuration/entry-context/
+      'context',
+      // https://webpack.js.org/configuration/mode/
+      'mode',
+      // https://webpack.js.org/configuration/devtool/
+      'devtool',
+      // https://webpack.js.org/configuration/target/
+      'target',
+      // https://webpack.js.org/configuration/watch/
+      'watch',
+      'watchOptions',
+      // https://webpack.js.org/configuration/externals/
+      'externals',
+      'externalsType',
+      'externalsPresets',
+      // https://webpack.js.org/configuration/stats/
+      'stats',
+      // https://webpack.js.org/configuration/experiments
+      'experiments',
+      // https://webpack.js.org/configuration/other-options
       'amd',
       'bail',
       'cache',
-      'context',
-      'devtool',
-      'externals',
+      'dependencies',
+      'ignoreWarnings',
       'loader',
-      'mode',
-      'name',
       'parallelism',
       'profile',
-      'recordsInputPath',
       'recordsPath',
+      'recordsInputPath',
       'recordsOutputPath',
-      'stats',
-      'target',
-      'watch',
-      'watchOptions',
+      'name',
+      'infrastructureLogging',
+      'snapshot',
     ]);
   }
 
@@ -116,9 +143,10 @@ module.exports = class extends ChainedMap {
 
   toConfig() {
     const entryPoints = this.entryPoints.entries() || {};
+    const baseConfig = this.entries() || {};
 
     return this.clean(
-      Object.assign(this.entries() || {}, {
+      Object.assign(baseConfig, {
         node: this.node.entries(),
         output: this.output.entries(),
         resolve: this.resolve.toConfig(),
