@@ -12,9 +12,9 @@ module.exports = class extends ChainedMap {
       'before',
       'bonjour',
       'clientLogLevel',
-      'color',
       'compress',
       'contentBase',
+      'contentBasePublicPath',
       'disableHostCheck',
       'filename',
       'headers',
@@ -25,11 +25,14 @@ module.exports = class extends ChainedMap {
       'http2',
       'https',
       'index',
-      'info',
+      'injectClient',
+      'injectHot',
       'inline',
       'lazy',
+      'liveReload',
       'mimeTypes',
       'noInfo',
+      'onListening',
       'open',
       'openPage',
       'overlay',
@@ -41,6 +44,7 @@ module.exports = class extends ChainedMap {
       'public',
       'publicPath',
       'quiet',
+      'serveIndex',
       'setup',
       'socket',
       'sockHost',
@@ -49,6 +53,7 @@ module.exports = class extends ChainedMap {
       'staticOptions',
       'stats',
       'stdin',
+      'transportMode',
       'useLocalIp',
       'watchContentBase',
       'watchOptions',
