@@ -5,36 +5,50 @@ module.exports = class extends ChainedMap {
     super(parent);
     this.extend([
       'auxiliaryComment',
-      'chunkCallbackName',
+      'charset',
       'chunkFilename',
       'chunkLoadTimeout',
+      'chunkLoadingGlobal',
+      'chunkLoading',
+      'chunkFormat',
+      'enabledChunkLoadingTypes',
       'crossOriginLoading',
       'devtoolFallbackModuleFilenameTemplate',
-      'devtoolLineToLine',
       'devtoolModuleFilenameTemplate',
       'devtoolNamespace',
       'filename',
-      'futureEmitAssets',
+      'assetModuleFilename',
       'globalObject',
+      'uniqueName',
       'hashDigest',
       'hashDigestLength',
       'hashFunction',
       'hashSalt',
       'hotUpdateChunkFilename',
-      'hotUpdateFunction',
+      'hotUpdateGlobal',
       'hotUpdateMainFilename',
-      'jsonpFunction',
       'library',
       'libraryExport',
       'libraryTarget',
+      'importFunctionName',
       'path',
       'pathinfo',
       'publicPath',
+      'scriptType',
       'sourceMapFilename',
       'sourcePrefix',
+      'strictModuleErrorHandling',
       'strictModuleExceptionHandling',
       'umdNamedDefine',
-      'webassemblyModuleFilename',
+      'workerChunkLoading',
+      'enabledLibraryTypes',
+      'environment',
+      'compareBeforeEmit',
+      'wasmLoading',
+      'enabledWasmLoadingTypes',
+      'iife',
+      'module',
+      'clean',
     ]);
   }
 };
