@@ -1,6 +1,6 @@
-const ChainedMap = require('./ChainedMap');
+const ChainedValueMap = require('./ChainedValueMap');
 
-module.exports = class extends ChainedMap {
+module.exports = class extends ChainedValueMap {
   constructor(parent) {
     super(parent);
     this.extend(['assetFilter', 'hints', 'maxAssetSize', 'maxEntrypointSize']);
