@@ -4,12 +4,14 @@ const ChainedSet = require('./ChainedSet');
 module.exports = class extends Resolve {
   constructor(parent) {
     super(parent);
+    this.modules = new ChainedSet(this);
     this.moduleExtensions = new ChainedSet(this);
     this.packageMains = new ChainedSet(this);
   }
 
   toConfig() {
     return this.clean({
+      modules: this.modules.values(),
       moduleExtensions: this.moduleExtensions.values(),
       packageMains: this.packageMains.values(),
       ...super.toConfig(),
@@ -17,7 +19,7 @@ module.exports = class extends Resolve {
   }
 
   merge(obj, omit = []) {
-    const omissions = ['moduleExtensions', 'packageMains'];
+    const omissions = ['modules', 'moduleExtensions', 'packageMains'];
 
     omissions.forEach((key) => {
       if (!omit.includes(key) && key in obj) {
